+++ src/main/java/com/zea7ot/leetcode/logs.txt
@@ -23,12 +23,12 @@ review
 20200820:
 new
 3: 0904
-4: 
+4:
 5: 
 
 review
 3: 0497
-4: 0159, 1032
+4: 0159, 0930, 1032
 5: 0862
 
 20200819:
