+++ src/main/java/com/zea7ot/leetcode/tags/algorithm/dynamic_programming/prefix_sums.txt
@@ -23,6 +23,7 @@ References:
 
 4   0410    https://leetcode.com/problems/split-array-largest-sum/
 4   0689    https://leetcode.com/problems/maximum-sum-of-3-non-overlapping-subarrays/
+4   0930    https://leetcode.com/problems/binary-subarrays-with-sum/
 4   1124    https://leetcode.com/problems/longest-well-performing-interval/
 4   1292    https://leetcode.com/problems/maximum-side-length-of-a-square-with-sum-less-than-or-equal-to-threshold/
 4   1371    https://leetcode.com/problems/find-the-longest-substring-containing-vowels-in-even-counts/
@@ -31,6 +32,17 @@ References:
 5   0862    https://leetcode.com/problems/shortest-subarray-with-sum-at-least-k/
 
 6   1000    https://leetcode.com/problems/minimum-cost-to-merge-stones/
+
+
+
+# how many
+
+4   0930    https://leetcode.com/problems/binary-subarrays-with-sum/
+4   0992    https://leetcode.com/problems/subarrays-with-k-different-integers/
+
+
+
+# shortest / longest
