+++ src/main/java/com/zea7ot/leetcode/tags/algorithm/dynamic_programming/summary.txt
@@ -90,6 +90,7 @@ This serves the total list of problems related to Dynamic Programming, regardles
 4   0894    https://leetcode.com/problems/all-possible-full-binary-trees/
 4   0907    https://leetcode.com/problems/sum-of-subarray-minimums/
 4   0918    https://leetcode.com/problems/maximum-sum-circular-subarray/
+4   0930    https://leetcode.com/problems/binary-subarrays-with-sum/
 4   0956    https://leetcode.com/problems/tallest-billboard/
 4   0983    https://leetcode.com/problems/minimum-cost-for-tickets/
 4   1027    https://leetcode.com/problems/longest-arithmetic-sequence/
