+++ src/main/java/com/zea7ot/leetcode/tags/data_structure/array/dynamic_programming/prefix_sums.txt
@@ -18,6 +18,7 @@ References:
 3   0974    https://leetcode.com/problems/subarray-sums-divisible-by-k/
 3   1031    https://leetcode.com/problems/maximum-sum-of-two-non-overlapping-subarrays/
 
+4   0930    https://leetcode.com/problems/binary-subarrays-with-sum/
 4   1124    https://leetcode.com/problems/longest-well-performing-interval/
 4   1371    https://leetcode.com/problems/find-the-longest-substring-containing-vowels-in-even-counts/
 4   1477    https://leetcode.com/problems/find-two-non-overlapping-sub-arrays-each-with-target-sum/
@@ -25,6 +26,7 @@ References:
 5   0862    https://leetcode.com/problems/shortest-subarray-with-sum-at-least-k/
 
 6   1000    https://leetcode.com/problems/minimum-cost-to-merge-stones/
+
