+++ src/main/java/com/zea7ot/leetcode/tags/data_structure/array/summary.txt
@@ -192,6 +192,7 @@ This file serves as the total list of problems related to Array, regardless of a
 4   0889    https://leetcode.com/problems/construct-binary-tree-from-preorder-and-postorder-traversal/
 4   0916    https://leetcode.com/problems/word-subsets/
 4   0918    https://leetcode.com/problems/maximum-sum-circular-subarray/
+4   0930    https://leetcode.com/problems/binary-subarrays-with-sum/
 4   0947    https://leetcode.com/problems/most-stones-removed-with-same-row-or-column/
 4   0956    https://leetcode.com/problems/tallest-billboard/
 4   0980    https://leetcode.com/problems/unique-paths-iii/
