+++ src/main/java/com/zea7ot/leetcode/tags/technique/n_pointers/2_pointers/summary.txt
@@ -39,6 +39,7 @@ References:
 4   0287    https://leetcode.com/problems/find-the-duplicate-number/
 4   0632    https://leetcode.com/problems/smallest-range-covering-elements-from-k-lists/
 4   0689    https://leetcode.com/problems/maximum-sum-of-3-non-overlapping-subarrays/
+4   0930    https://leetcode.com/problems/binary-subarrays-with-sum/
 4   1537    https://leetcode.com/problems/get-the-maximum-score/
