+++ src/main/java/com/zea7ot/leetcode/tags/technique/n_pointers/hi_lo_pointers.txt
@@ -17,6 +17,7 @@ References:
 3   1498    https://leetcode.com/problems/number-of-subsequences-that-satisfy-the-given-sum-condition/
 
 4   0159    https://leetcode.com/problems/longest-substring-with-at-most-two-distinct-characters/
+4   0930    https://leetcode.com/problems/binary-subarrays-with-sum/
